[FIX] Deal with a weird astropy bug

+++ construct_hexagon_1kpc.py
@@ -56,23 +56,23 @@ def gen_tessell_mega_table(
 
     # record metadata
     t.meta['GALAXY'] = str(gal_params['name'])
-    t.meta['DIST_MPC'] = gal_params['dist_Mpc']
-    t.meta['RA_DEG'] = gal_params['ra_deg']
-    t.meta['DEC_DEG'] = gal_params['dec_deg']
-    t.meta['INCL_DEG'] = gal_params['incl_deg']
-    t.meta['PA_DEG'] = gal_params['posang_deg']
-    t.meta['LOGMSTAR'] = np.log10(gal_params['Mstar_Msun'])
-    t.meta['R25_KPC'] = (
+    t.meta['DIST_MPC'] = np.float(gal_params['dist_Mpc'])
+    t.meta['RA_DEG'] = np.float(gal_params['ra_deg'])
+    t.meta['DEC_DEG'] = np.float(gal_params['dec_deg'])
+    t.meta['INCL_DEG'] = np.float(gal_params['incl_deg'])
+    t.meta['PA_DEG'] = np.float(gal_params['posang_deg'])
+    t.meta['LOGMSTAR'] = np.float(np.log10(gal_params['Mstar_Msun']))
+    t.meta['R25_KPC'] = np.float((
         (gal_params['R25_arcsec'] * u.arcsec).to('rad').value *
-        gal_params['dist_Mpc'] * u.Mpc).to('kpc').value
-    t.meta['RDISKKPC'] = (
+        gal_params['dist_Mpc'] * u.Mpc).to('kpc').value)
+    t.meta['RDISKKPC'] = np.float((
         (gal_params['Rstar_arcsec'] * u.arcsec).to('rad').value *
-        gal_params['dist_Mpc'] * u.Mpc).to('kpc').value
-    t.meta['CO_R21'] = phys_params['CO_R21']
-    t.meta['H_MOL_PC'] = phys_params['CO_full_height']
-    t.meta['ABUN_SUN'] = phys_params['abundance_solar']
+        gal_params['dist_Mpc'] * u.Mpc).to('kpc').value)
+    t.meta['CO_R21'] = np.float(phys_params['CO_R21'])
+    t.meta['H_MOL_PC'] = np.float(phys_params['CO_full_height'])
+    t.meta['ABUN_SUN'] = np.float(phys_params['abundance_solar'])
     t.meta['TBLNOTE'] = str(note)
-    t.meta['VERSION'] = float(version)
+    t.meta['VERSION'] = np.float(version)
 
     # output
     if writefile:
